# Conflicts:
#	beautify_api/8.0.0/aspnet-core/src/BanHangBeautify.EntityFrameworkCore/Migrations/SPADbContextModelSnapshot.cs

+++ beautify_api/8.0.0/aspnet-core/src/BanHangBeautify.Application/ChietKhau/ChietKhauDichVu/Dto/ChietKhauDichVuDto.cs
@@ -14,7 +14,7 @@ namespace BanHangBeautify.ChietKhau.ChietKhauDichVu.Dto
         public Guid IdNhanVien { set; get; }
         public Guid IdDonViQuiDoi { set; get; }
         public byte? LoaiChietKhau { set; get; } = 1;
-        public float? GiaTri { set; get; } = 0;
+        public double? GiaTri { set; get; } = 0;
         public bool? LaPhanTram { set; get; } = true;
 
     }
