+++ beautify_api/8.0.0/aspnet-core/src/BanHangBeautify.Application/ChietKhau/ChietKhauHoaDon/Dto/ChietKhauHoaDonDto.cs
@@ -11,7 +11,7 @@ namespace BanHangBeautify.ChietKhau.ChietKhauHoaDon.Dto
         public Guid Id{ set; get; }
         public Guid IdChiNhanh{ set; get; }
         public byte LoaiChietKhau{ set; get; }
-        public decimal GiaTriChietKhau{ set; get; }
+        public double GiaTriChietKhau { set; get; }
         public List<string> ChungTuApDung{ set; get; }
     }
 }
