+++ beautify_api/8.0.0/aspnet-core/src/BanHangBeautify.Application/HangHoa/HangHoa/Dto/HangHoaDto.cs
@@ -17,7 +17,7 @@ namespace BanHangBeautify.HangHoa.HangHoa.Dto
         public float? SoPhutThucHien { get; set; }
         public int TrangThai { get; set; }
         public int IdLoaiHangHoa { get; set; }
-        public float? GiaBan { get; set; }
+        public double? GiaBan { get; set; }
         public string MoTa { get; set; }
         public string TenLoaiHangHoa { get; set; }
         public string TxtTrangThaiHang { get; set; }
