+++ beautify_api/8.0.0/aspnet-core/src/BanHangBeautify.Application/HoaDon/NhanVienThucHien/Dto/NhanVienThucHienDto.cs
@@ -12,10 +12,10 @@ namespace BanHangBeautify.HoaDon.NhanVienThucHien.Dto
         public Guid? IdHoaDon { set; get; }
         public Guid? IdChiTietHoaDon { set; get; }
         public Guid? IdQuyHoaDon { set; get; }
-        public float? PTChietKhau { set; get; }
-        public float? TienChietKhau { set; get; }
+        public double? PTChietKhau { set; get; }
+        public double? TienChietKhau { set; get; }
         public bool? ChiaDeuChietKhau { set; get; } = false;
-        public float? HeSo { set; get; } = 1;
+        public double? HeSo { set; get; } = 1;
         public bool? TinhHoaHongTruocCK { set; get; } = false;
         public byte? LoaiChietKhau { set; get; } = 1;
     }
