+++ beautify_api/8.0.0/aspnet-core/src/BanHangBeautify.Application/Quy/DM_QuyHoaDon/Dto/CreateOrEditQuyHoaDonDto.cs
@@ -15,7 +15,7 @@ namespace BanHangBeautify.Quy.DM_QuyHoaDon.Dto
         public int IdLoaiChungTu{set;get;}
         public string MaHoaDon{set;get;}
         public DateTime NgayLapHoaDon{set;get;}
-        public float? TongTienThu { set; get; } = 0;
+        public double? TongTienThu { set; get; } = 0;
         public string NoiDungThu{set;get;}
         public bool HachToanKinhDoanh { set; get; }
         public bool? IsDelete { set; get; } = false;
