+++ beautify_api/8.0.0/aspnet-core/src/BanHangBeautify.Application/Quy/QuyHoaDonChiTiet/Dto/CreateOrEditQuyHoaDonCTDto.cs
@@ -16,11 +16,11 @@ namespace BanHangBeautify.Quy.QuyHoaDonChiTiet.Dto
         public Guid? IdNhanVien{set;get;}
         public Guid? IdTaiKhoanNganHang{set;get;}
         public Guid? IdKhoanThuChi{set;get;}
-        public float? LaPTChiPhiNganHang { get; set; } = 0;
-        public float? ChiPhiNganHang { get; set; } = 0;
-        public float? ThuPhiTienGui { get; set; } = 0;
-        public float? DiemThanhToan { get; set; } = 0;
+        public double? LaPTChiPhiNganHang { get; set; } = 0;
+        public double? ChiPhiNganHang { get; set; } = 0;
+        public double? ThuPhiTienGui { get; set; } = 0;
+        public double? DiemThanhToan { get; set; } = 0;
         public byte HinhThucThanhToan { get; set; } = 1;
-        public float? TienThu { get; set; } = 0;
+        public double? TienThu { get; set; } = 0;
     }
 }
